Modify keras code and add loss & iou plot

+++ unet/metrics.py
@@ -9,7 +9,7 @@ from keras import backend as K
 def mean_iou(y_true, y_pred):
     prec = []
     for t in np.arange(0.5, 1.0, 0.05):
-        y_pred_ = tf.to_int32(y_pred > t)
+        y_pred_ = tf.cast(y_pred > t, tf.int32)
         score, up_opt = tf.metrics.mean_iou(y_true, y_pred_, 2)
         K.get_session().run(tf.local_variables_initializer())
         with tf.control_dependencies([up_opt]):
